+++ apps/web/app/future/video/meeting-ended/[uid]/page.tsx
@@ -2,9 +2,10 @@ import { withAppDirSsr } from "app/WithAppDirSsr";
 import { _generateMetadata } from "app/_utils";
 import { WithLayout } from "app/layoutHOC";
 
+import { getServerSideProps } from "@lib/video/meeting-ended/[uid]/getServerSideProps";
+
 import type { PageProps } from "~/videos/views/videos-meeting-ended-single-view";
 import MeetingEnded from "~/videos/views/videos-meeting-ended-single-view";
-import { getServerSideProps } from "~/videos/views/videos-meeting-ended-single-view.getServerSideProps";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
