+++ apps/web/pages/video/[uid].tsx
@@ -2,7 +2,7 @@ import PageWrapper from "@components/PageWrapper";
 
 import VideosSingleView, { type PageProps } from "~/videos/views/videos-single-view";
 
-export { getServerSideProps } from "~/videos/views/videos-single-view.getServerSideProps";
+export { getServerSideProps } from "@lib/video/[uid]/getServerSideProps";
 
 const VideosSinglePage = (props: PageProps) => <VideosSingleView {...props} />;
