+++ apps/web/pages/video/meeting-not-started/[uid].tsx
@@ -2,7 +2,7 @@ import PageWrapper from "@components/PageWrapper";
 
 import MeetingNotStarted, { type PageProps } from "~/videos/views/videos-meeting-not-started-single-view";
 
-export { getServerSideProps } from "~/videos/views/videos-meeting-not-started-single-view.getServerSideProps";
+export { getServerSideProps } from "@lib/video/meeting-not-started/[uid]/getServerSideProps";
 
 const MeetingNotStartedPage = (props: PageProps) => <MeetingNotStarted {...props} />;
